requirements.txt
flask
pandas
python-dotenv
locale
datetime
mysql-connector-python